move some enums
